requirements.txt
Django>=4.2,<5.0
psycopg2-binary
deep_translator
djangorestframework
djangorestframework-simplejwt
django-filter
Pillow
django-import-export